added group name validation to prevent path operations causing crashes
when a path with a character that is invalid is passed as a group name
it may cause the program to panic, such as `/group` or `group.` or `..`
All these group names make so that the group resolution becomes erroneous
and thus, it is undesirable. Instead of just letting it pass, we now
filter it out and tell the user those are invalid.