fixed Dotfile erroneously assigning group to Configs/Hooks/Secrets root dir
when Dotfile::try_new was being called it would accept .config/dotfiles/Configs
as a valid dotfile path even though that's false. This would result in
the groupis command giving a `Configs` group.

The Dotfile was also wrongly being used just to get an iterator in
the SymlinkHandler validation logic, so that has been removed as well.