fixed `get_related_conditional_groups` returning None on target groups
When a group like `test_linux` was used the related cond groups
would just return None, this would result in `tuckr status <group>`
not printing anything. Whereas if `test` was using it would print
`vec!["test"]` so this unifies the behavior and cuts on the amount of
complexity needed to extract status information.